refactor(frontend): explicit clear semantics in topic edit; extract SeasonEpisodePicker
